Custom authorization policies (#413)
* p-o-c custom policies

* add authorization to API routes

* move all authorization to controllers

+++ src/Http/Api/Controllers/EmailLists/EmailListsController.php
@@ -2,6 +2,7 @@
 
 namespace Spatie\Mailcoach\Http\Api\Controllers\EmailLists;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Spatie\Mailcoach\Domain\Campaign\Actions\EmailLists\UpdateEmailListAction;
 use Spatie\Mailcoach\Domain\Campaign\Models\EmailList;
 use Spatie\Mailcoach\Domain\Shared\Traits\UsesMailcoachModels;
@@ -12,20 +13,28 @@ use Spatie\Mailcoach\Http\App\Queries\EmailListQuery;
 
 class EmailListsController
 {
-    use RespondsToApiRequests, UsesMailcoachModels;
+    use AuthorizesRequests,
+        RespondsToApiRequests,
+        UsesMailcoachModels;
 
     public function index(EmailListQuery $emailLists)
     {
+        $this->authorize("viewAny", EmailList::class);
+
         return EmailListResource::collection($emailLists->paginate());
     }
 
     public function show(EmailList $emailList)
     {
+        $this->authorize("view", $emailList);
+
         return new EmailListResource($emailList);
     }
 
     public function store(EmailListRequest $request, UpdateEmailListAction $updateEmailListAction)
     {
+        $this->authorize("create", EmailList::class);
+
         $emailListClass = $this->getEmailListClass();
 
         $emailList = new $emailListClass;
@@ -37,6 +46,8 @@ class EmailListsController
 
     public function update(EmailListRequest $request, EmailList $emailList, UpdateEmailListAction $updateEmailListAction)
     {
+        $this->authorize("update", $emailList);
+
         $emailList = $updateEmailListAction->execute($emailList, $request);
 
         return new EmailListResource($emailList);
@@ -44,6 +55,8 @@ class EmailListsController
 
     public function destroy(EmailList $emailList)
     {
+        $this->authorize("delete", $emailList);
+
         $emailList->delete();
 
         return $this->respondOk();
