+++ src/Http/Api/Controllers/SubscriberImports/SubscriberImportsController.php
@@ -2,19 +2,26 @@
 
 namespace Spatie\Mailcoach\Http\Api\Controllers\SubscriberImports;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Response;
 use Spatie\Mailcoach\Domain\Campaign\Enums\SubscriberImportStatus;
+use Spatie\Mailcoach\Domain\Campaign\Models\EmailList;
 use Spatie\Mailcoach\Domain\Campaign\Models\SubscriberImport;
+use Spatie\Mailcoach\Domain\Shared\Traits\UsesMailcoachModels;
 use Spatie\Mailcoach\Http\Api\Controllers\Concerns\RespondsToApiRequests;
 use Spatie\Mailcoach\Http\Api\Requests\SubscriberImportRequest;
 use Spatie\Mailcoach\Http\Api\Resources\SubscriberImportResource;
 
 class SubscriberImportsController
 {
+    use AuthorizesRequests;
+    use UsesMailcoachModels;
     use RespondsToApiRequests;
 
     public function index()
     {
+        $this->authorize("viewAny", EmailList::class);
+
         $subscribersImport = SubscriberImport::query()->paginate();
 
         return SubscriberImportResource::collection($subscribersImport);
@@ -22,11 +29,15 @@ class SubscriberImportsController
 
     public function show(SubscriberImport $subscriberImport)
     {
+        $this->authorize("view", $subscriberImport->emailList);
+
         return new SubscriberImportResource($subscriberImport);
     }
 
     public function store(SubscriberImportRequest $request)
     {
+        $this->authorize("update", self::getEmailListClass()::findOrFail($request->email_list_id));
+
         $attributes = array_merge($request->validated(), ['status' => SubscriberImportStatus::DRAFT]);
 
         $subscriberImport = SubscriberImport::create($attributes);
@@ -36,6 +47,8 @@ class SubscriberImportsController
 
     public function update(SubscriberImportRequest $request, SubscriberImport $subscriberImport)
     {
+        $this->authorize("update", $subscriberImport->emailList);
+
         if ($subscriberImport->status !== SubscriberImportStatus::DRAFT) {
             abort(Response::HTTP_UNPROCESSABLE_ENTITY, 'Cannot update a non-draft import.');
         }
@@ -47,6 +60,8 @@ class SubscriberImportsController
 
     public function destroy(SubscriberImport $subscriberImport)
     {
+        $this->authorize("update", $subscriberImport->emailList);
+
         $subscriberImport->delete();
 
         return $this->respondOk();
